fx quant: do not observe bias on F.linear
Summary:

Ensures that linear bias is not observed in a `F.linear` call. This should
be a small speedup in PTQ, and will change numerics (in a good way) for
QAT if someone is using `F.linear`.

Note: the implementation is slightly more verbose compared to conv
because bias is a keyword argument in Linear.

Test Plan:

```
python test/test_quantization.py TestQuantizeFxOps.test_linear_functional_bias_not_observed
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: b43e35910276c5f43354d0d9962f7df7b8221c79
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49628